Added foreach to loop through grades
Created an if-else loop inside of the foreach loop iterating through grades[]. For each value, the console will return the letter grade corresponding to the number.

I had i>100 or i<0 as an elseif instead of an else; the code in the else section should be impossible if this code works (im doing this in notepad).