Logic Error in "average" + Console edits
Edited console messages for aesthetics. Shortened the "out of range" response to be closer in length to the other responses.

Fixed a logic error I had made when calculating average. I noticed it when I changed a value in grades[] to be way out of range. Before, it added up all the values from grades[]. If you had any values in grades[] outside of 0-100, which is a value you aren't supposed to have, you can get an average outside of that range. I corrected this by making a separate variable  "values" to iterate thru grades[] and increase by one for each value in the 0-100 range. I also made average a float since you get decimal values with actual grades.